refactor: extract hook and session command handlers
